requirements.txt
streamlit==1.23.1
PyYAML==6.0
Pillow==9.5.0
pandas==2.0.1
wordcloud==1.9.2
seaborn==0.12.2
matplotlib==3.7.1 
azure-ai-textanalytics==5.3.0
azure-core==1.27.0
numpy==1.24.3
openai==0.27.7
azure-ai-translation-text==1.0.0b1
plotly==5.17.0
pyxlsb==1.0.10
openpyxl==3.1.2
xlsxwriter==3.1.6